+++ src/lang/r.js
@@ -1,0 +1,21 @@
+// ----------------------------------------------------------------------
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+// --
+// Copyright 2020 Andi Dittrich <https://andidittrich.de>
+// ----------------------------------------------------------------------
+
+// Generic Rules/Regex
+import {generic} from './generic';
+
+// r language
+// Author: [Andi Dittrich]
+// --
+export class r extends generic {
+
+    // language aliases
+    static alias(){
+        return [];
+    }
+}